04 Complete

+++ 04.Poisson_equation/src/step.h
@@ -1,4 +1,4 @@
 #include "data_types.h"
 
-int step(Parameters parameters, Cart_Data cart_data, Local_Data local_data,
+double step(Parameters parameters, Cart_Data cart_data, Local_Data local_data,
          MPI_Datatype column_datatype);
